moved buttons component folder out of navigation component folder
