+++ med-booking-frontend/src/Components/Navigation/AdminNav.tsx
@@ -2,8 +2,8 @@ import { Button } from "@mui/material";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
-import { AdminLoginButton } from "./buttons/AdminLoginButton";
-import { AdminLogoutButton } from "./buttons/AdminLogoutButton";
+import { AdminLogoutButton } from "../buttons/AdminLogoutButton";
+import { AdminLoginButton } from "../buttons/AdminLoginButton";
 
 export default function AdminNav(props: NavigationProps) {
   const { isUserAuthenticated } = props;
